+++ src/headers/ModItem.hpp
@@ -12,15 +12,15 @@ class ModItem : public CCNode {
 protected:
     bool m_favorite = false; // If the mod is favorite
 
-    bool m_geodeTheme = false; // Make sure visuals go with geode theme
+    bool m_geodeTheme = false; // Make sure visuals go with Geode theme
     bool m_heartIcons = false; // Hearts UI mode
 
     Mod* m_mod; // Fetched mod
     Mod* m_thisMod = getMod(); // Get this mod
 
-    CCMenuItemSpriteExtra* m_favButton = nullptr;
-    FavoritesPopup* m_parentPopup = nullptr;
-    CCScale9Sprite* m_backgroundSprite = nullptr;
+    CCMenuItemSpriteExtra* m_favButton = nullptr; // Favorite button
+    FavoritesPopup* m_parentPopup = nullptr; // To access parent methods
+    CCScale9Sprite* m_backgroundSprite = nullptr; // Background theme
 
     void onModDesc(CCObject*);
     void onViewMod(CCObject*);