+++ src/headers/src/FavoritesPopup.cpp
@@ -55,7 +55,7 @@ bool FavoritesPopup::setup() {
     auto [widthCS, heightCS] = contentSize;
 
     // Create scroll size
-    auto scrollSize = CCSize{ widthCS - 20.f, heightCS - 90.f };
+    auto scrollSize = CCSize{ widthCS - 17.5f, heightCS - 80.f };
 
     // for buttons to work
     m_overlayMenu = CCMenu::create();
@@ -71,7 +71,7 @@ bool FavoritesPopup::setup() {
     // Create search input
     m_searchInput = TextInput::create(widthCS - 135.f, "Search Mods", "bigFont.fnt");
     m_searchInput->setID("search-box");
-    m_searchInput->setPosition({ widthCS / 2.f - 20.f, heightCS - 60.f });
+    m_searchInput->setPosition({ widthCS / 2.f - 20.f, heightCS - 50.f });
     m_searchInput->setDelegate(this);
     m_searchInput->setMaxCharCount(50);
 
@@ -116,7 +116,7 @@ bool FavoritesPopup::setup() {
         menu_selector(FavoritesPopup::onFavoritesOnlyToggle)
     );
     m_favoritesOnlyToggle->setID("show-favorites-only");
-    m_favoritesOnlyToggle->setPosition({ widthCS - 70.f, heightCS - 60.f });
+    m_favoritesOnlyToggle->setPosition({ widthCS - 70.f, heightCS - 50.f });
 
     checkboxMenu->addChild(m_favoritesOnlyToggle);
 
@@ -134,20 +134,20 @@ bool FavoritesPopup::setup() {
         menu_selector(FavoritesPopup::onHideFavoritesToggle)
     );
     m_hideFavoritesToggle->setID("hide-all-favorites");
-    m_hideFavoritesToggle->setPosition({ widthCS - 30.f, heightCS - 60.f });
+    m_hideFavoritesToggle->setPosition({ widthCS - 30.f, heightCS - 50.f });
 
     checkboxMenu->addChild(m_hideFavoritesToggle);
 
     // Create icon above favorites only toggle
     auto starOnIcon = CCSprite::createWithSpriteFrameName(m_heartIcons ? "gj_heartOn_001.png" : "GJ_starsIcon_001.png");
     starOnIcon->setID("favorites-icon");
-    starOnIcon->setPosition({ widthCS - 70.f, heightCS - 40.f });
+    starOnIcon->setPosition({ m_favoritesOnlyToggle->getPositionX(), m_favoritesOnlyToggle->getPositionY() + 20.f });
     starOnIcon->setScale(m_heartIcons ? 0.375f : 0.625f);
 
     // Create icon above hide favorites toggle
     auto starOffIcon = CCSprite::createWithSpriteFrameName(m_heartIcons ? "gj_heartOff_001.png" : "GJ_starsIcon_gray_001.png");
     starOffIcon->setID("non-favorites-icon");
-    starOffIcon->setPosition({ widthCS - 30.f, heightCS - 40.f });
+    starOffIcon->setPosition({ m_hideFavoritesToggle->getPositionX(), m_hideFavoritesToggle->getPositionY() + 20.f });
     starOffIcon->setScale(m_heartIcons ? 0.375f : 0.625f);
 
     m_mainLayer->addChild(starOnIcon);
@@ -176,7 +176,7 @@ bool FavoritesPopup::setup() {
     // Create "No mods found" label (initially hidden)
     m_noModsLabel = CCLabelBMFont::create("No mods found :(", "bigFont.fnt");
     m_noModsLabel->setID("no-mods-label");
-    m_noModsLabel->setPosition({ widthCS / 2.f, heightCS / 2.f - 35.f });
+    m_noModsLabel->setPosition({ widthCS / 2.f, heightCS / 2.f - 30.f });
     m_noModsLabel->setScale(0.5f);
     m_noModsLabel->setColor({ 125, 125, 125 });
     m_noModsLabel->setVisible(false);
@@ -188,7 +188,7 @@ bool FavoritesPopup::setup() {
     scrollBG->setContentSize(scrollSize);
     scrollBG->setAnchorPoint({ 0.5, 0.5 });
     scrollBG->ignoreAnchorPointForPosition(false);
-    scrollBG->setPosition({ widthCS / 2.f, (heightCS / 2.f) - 35.f });
+    scrollBG->setPosition({ widthCS / 2.f, (heightCS / 2.f) - 30.f });
     scrollBG->setColor({ 0, 0, 0 });
     scrollBG->setOpacity(100);